requirements.txt
PyQt5>=5.15.0
google-api-python-client>=2.0.0
google-auth>=2.0.0
google-auth-oauthlib>=0.5.0
google-auth-httplib2>=0.1.0
httplib2>=0.20.0
oauth2client>=4.1.3
